% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog_slow.R
\name{mylog_slow}
\alias{mylog_slow}
\title{My log function}
\usage{
mylog_slow(x)
}
\arguments{
\item{x}{a numerical vector of length n}
}
\value{
a vector of log(x) of length n
}
\description{
My log function
}
\examples{
mylog(c(10, exp(2), 1))
}
